requirements.txt
rich==11.2.0
python-dotenv~=0.19.2
setuptools==60.8.2
PyYAML~=6.0